cookies working
still using fakeredis and only one person allowed at a time being logged in

+++ src/main/java/tukano/impl/JavaBlobs.java
@@ -13,71 +13,104 @@ import tukano.impl.storage.BlobStorage;
 import tukano.impl.storage.FilesystemStorage;
 import utils.Hash;
 import utils.Hex;
+import utils.auth.Authentication;
 
 public class JavaBlobs implements Blobs {
-	
+
 	private static Blobs instance;
 	private static Logger Log = Logger.getLogger(JavaBlobs.class.getName());
 
 	public String baseURI;
 	private BlobStorage storage;
-	
+
 	synchronized public static Blobs getInstance() {
-		if( instance == null )
+		if (instance == null)
 			instance = new JavaBlobs();
 		return instance;
 	}
-	
+
 	private JavaBlobs() {
 		storage = new FilesystemStorage();
 		baseURI = String.format("%s/%s/", TukanoRestServer.serverURI, Blobs.NAME);
 	}
-	
+
 	@Override
 	public Result<Void> upload(String blobId, byte[] bytes, String token) {
-		Log.info(() -> format("upload : blobId = %s, sha256 = %s, token = %s\n", blobId, Hex.of(Hash.sha256(bytes)), token));
+		if (!validateSession(blobId)) {
+			return error(FORBIDDEN);
+		}
+
+		Log.info(() -> format("upload : blobId = %s, sha256 = %s, token = %s\n", blobId, Hex.of(Hash.sha256(bytes)),
+				token));
 
 		if (!validBlobId(blobId, token))
 			return error(FORBIDDEN);
 
-		return storage.write( toPath( blobId ), bytes);
+		return storage.write(toPath(blobId), bytes);
 	}
 
 	@Override
 	public Result<byte[]> download(String blobId, String token) {
+		if (!validateSession(blobId)) {
+			return error(FORBIDDEN);
+		}
+
 		Log.info(() -> format("download : blobId = %s, token=%s\n", blobId, token));
 
-		if( ! validBlobId( blobId, token ) )
+		if (!validBlobId(blobId, token))
 			return error(FORBIDDEN);
 
-		return storage.read( toPath( blobId ) );
+		return storage.read(toPath(blobId));
 	}
 
 	@Override
 	public Result<Void> delete(String blobId, String token) {
+		if (!validateSession(blobId)) {
+			return error(FORBIDDEN);
+		}
+
 		Log.info(() -> format("delete : blobId = %s, token=%s\n", blobId, token));
-	
-		if( ! validBlobId( blobId, token ) )
+
+		if (!validBlobId(blobId, token))
 			return error(FORBIDDEN);
 
-		return storage.delete( toPath(blobId));
+		return storage.delete(toPath(blobId));
 	}
-	
+
 	@Override
 	public Result<Void> deleteAllBlobs(String userId, String token) {
+		try {
+			Authentication.validateSession(userId);
+		} catch (Exception e) {
+			return error(FORBIDDEN);
+		}
+
 		Log.info(() -> format("deleteAllBlobs : userId = %s, token=%s\n", userId, token));
 
-		if( ! Token.isValid( token, userId ) )
+		if (!Token.isValid(token, userId))
 			return error(FORBIDDEN);
-		
-		return storage.delete( toPath(userId));
+
+		return storage.delete(toPath(userId));
 	}
-	
-	private boolean validBlobId(String blobId, String token) {		
+
+	private boolean validBlobId(String blobId, String token) {
 		return Token.isValid(token, blobId);
 	}
 
 	private String toPath(String blobId) {
 		return blobId.replace("+", "/");
+	}
+
+	private boolean validateSession(String blobId) {
+		String tempBlobId = blobId.replace("+", "/");
+		Log.info(() -> format("VALIDATE SESSION1 : blobId = %s\n", tempBlobId));
+		String userId = tempBlobId.split("/")[0];
+		Log.info(() -> format("VALIDATE SESSION2 : userId = %s\n", userId));
+		try {
+			Authentication.validateSession(userId);
+			return true;
+		} catch (Exception e) {
+			return false;
+		}
 	}
 }
